backend/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
django-cors-headers==3.10.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

